+++ at_utils/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.0.9
+- Changed at_commons dependency from 3.0.11 to ^3.0.11
+## 3.0.8
+- at_commons version change for renaming NotifyDelete verb to NotifyRemove
+## 3.0.7
+- at_commons version change for Info and NoOp verb
+- at_commons version change for NotifyDelete verb
 ## 3.0.6
 - at_commons version change for AtTimeoutException
 ## 3.0.5